[level 1] Title: [1차] 비밀지도, Time: 4.39 ms, Memory: 83.7 MB -BaekjoonHub

+++ 프로그래머스/1/17681. ［1차］ 비밀지도/［1차］ 비밀지도.java
@@ -4,12 +4,12 @@ import java.util.stream.IntStream;
 class Solution {
     public String[] solution(int n, int[] arr1, int[] arr2) {
         return IntStream.range(0, n)
-            .mapToObj(i -> toSecretMap(Long.valueOf(arr1[i] | arr2[i]), n))
+            .mapToObj(i -> toSecretMap(arr1[i] | arr2[i], n))
             .toArray(String[]::new);
     }
     
-    private String toSecretMap(long value, int n) {
-        String str = Long.toBinaryString(value);
+    private String toSecretMap(int value, int n) {
+        String str = Integer.toBinaryString(value);
         StringBuilder sb = new StringBuilder();
         
         if(n - str.length() > 0) {
